+++ internal/workspace/web.go
@@ -13,6 +13,7 @@ import (
 	"github.com/leg100/otf/internal/organization"
 	"github.com/leg100/otf/internal/rbac"
 	"github.com/leg100/otf/internal/resource"
+	"github.com/leg100/otf/internal/state"
 	"github.com/leg100/otf/internal/vcsprovider"
 )
 
@@ -21,6 +22,7 @@ type (
 		html.Renderer
 		auth.TeamService
 		VCSProviderService
+		state.StateService
 
 		svc Service
 	}